Inline CKRepresentable conformances and unify metadata fields
- Move CKRepresentable extensions from separate files into their respective object files
- Replace dateFields on DateObject with metadata on IDObject that includes date fields, launch/user IDs, and version
- Simplify toRecord implementations by using the unified metadata dictionary
